+++ HISTORY.rst
@@ -10,6 +10,11 @@ This version allows users to calculate normalization scaling factors,
 Added
 *******
 * Added new parameter to normalization functions 'return_scaling_factors', that allows you to access the scaling factors calculated by RNAlysis.
+* Added parameter 'gzip_output' to CutAdapt adapter trimming (fastq.trim_adapters_single_end and fastq.trim_adapters_paired_end), allowing users to determine whether or not the output files will be gzipped.
+
+Changed
+*******
+* utAdapt adapter trimming (fastq.trim_adapters_single_end and fastq.trim_adapters_paired_end) now outputs non-gzipped files by default.
 
 Fixed
 *******